pyproject.toml
[project]
name = "qbittorrent-jobs"
version = "0.1.0"
description = "A script that finds potential hit and run torrents in qBittorrent and tags them."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "json5>=0.10.0",
    "python-dotenv>=1.0.1",
    "qbittorrent-api>=2025.2.0",
]

[dependency-groups]
dev = [
    "ruff==0.13.*",
]
